Low dense effort level

+++ src/place/placers/analytical/GradientPlacer.java
@@ -93,7 +93,7 @@ public abstract class GradientPlacer extends AnalyticalAndGradientPlacer {
         options.add(
                 O_OUTER_EFFORT_LEVEL_SPARSE,
                 "number of solve-legalize iterations for sparse designs",
-                new Integer(20));
+                new Integer(15));
         options.add(
                 O_OUTER_EFFORT_LEVEL_DENSE,
                 "number of solve-legalize iterations for dense designs",
